Reconcile decimal Quantity (master) with int Quantity (purchase module)
After rebasing the purchase module onto master's decimal-quantity work,
the boundaries where bill / challan quantities feed the purchase + stock
modules need to bridge decimal -> int:

- AwaitingPurchaseInvoiceDto.TotalQtyRemaining and
  PurchaseTemplateLineDto.SoldQty/PurchasedQty/RemainingQty switched to
  decimal so they survive the round-trip from InvoiceItem.Quantity
  (decimal(18,4)) without precision loss in the picker / template UI.
- FbrService stock-out path truncates to int when emitting a
  StockMovement, since StockMovement / PurchaseItem / GoodsReceiptItem /
  OpeningStockBalance are still int. TODO comment added to flip those
  to decimal as a follow-up so fractional sales propagate correctly
  through inventory accounting.

+++ DTOs/PurchaseBillDto.cs
@@ -147,7 +147,10 @@ namespace MyApp.Api.DTOs
         public int ClientId { get; set; }
         public string ClientName { get; set; } = "";
         public int LinesAwaiting { get; set; }
-        public int TotalQtyRemaining { get; set; }
+        // Decimal because the sum source (InvoiceItem.Quantity) is now
+        // decimal(18,4). PurchasedQty is still int on the purchase side,
+        // so the math (decimal - int) lands as decimal here.
+        public decimal TotalQtyRemaining { get; set; }
     }
 
     /// <summary>
@@ -188,9 +191,11 @@ namespace MyApp.Api.DTOs
         /// <summary>Combined description preview (first description or "first +N more").</summary>
         public string Description { get; set; } = "";
 
-        public int SoldQty { get; set; }
-        public int PurchasedQty { get; set; }
-        public int RemainingQty { get; set; }
+        // Decimal because the sale-side Quantity is now decimal(18,4).
+        // PurchasedQty stays consistent so subtraction lands as decimal.
+        public decimal SoldQty { get; set; }
+        public decimal PurchasedQty { get; set; }
+        public decimal RemainingQty { get; set; }
 
         /// <summary>UOM the operator typed at sale time (most common across the group, if any).</summary>
         public string? SaleUom { get; set; }
